Fixed suggestions from Scrutinizer
